\name{Simulating chromosomes with CNVs}
\title{ Simulating Chromosomes}
\description{
  Simulating chromosomes with clonal or subclonal copy number variations.
}
\usage{
genSimChroms(N, chr, loci, minLen, maxLen, datapath, chlens, save)
}
\arguments{
  \item{N}{ an integer; the number of simulated chromosomes to be generated. }
  \item{chr}{an integer; the total number of germline variants and
  somatic mutations to simulate in the tumor genome.}
  \item{loci}{an integer; the number of single nucleotide
    polymorphisms (SNPs) to simulate as meaurements made to estimate
    copy number.}
  \item{minLen}{an integer; the minimum length of simulated CNVs.}
  \item{maxLen}{an integer; the maximum length of simulated CNVs.}
  \item{chlens}{a vector of integers; the lengths of human chrsomosomes.}
  \item{save}{a logical value; indicates whether output should be saved.}
}
\details{
  Each simulated chromosome has a defined number of SNPs, or markers, as
  determined by the chromosome number and the chromosome lengths used. The
  positions of each CNV are generated using the minimum and maximum length 
  inputs and the fraction of cells with each CNV are generated by sampling
  from a dirichlet distribution. Other characteristics - such as whether
  there are one or two CNVs, and, if two, whether one is nested within
  the other, are determined by sampling values at random.
}
\value{
  The \code{genSimChroms} function returns a list of objects, each containing
  two objects itself: a simulated chromosome (\code{sim}) and the data
  generated therefrom (\code{dat}). If the option 'save' is set to 
  \code{TRUE}, the function will return no output and instead save
  the data for each simulation individually in the specified datapath and
  also save the list of simulated chromosomes (the underlying truth
  from which the data was generated) as a file 'chrSims.rda' in the 
  specified path.

  The \code{sim} component contains two objects:
  \describe{
    \item{\code{psi}}{the vector indicating the fraction of cells with
    each CNV;}
    \item{\code{clones}}{a list data frames describing
    each tumor clone's chromosome segments;}
  }
  
  The \code{clones} component contains a list of objects, each containing 
  a data frame:
  \describe{
    \item{\code{chr}}{the chromosome number;}
    \item{\code{start}}{the start point of each segment;}
    \item{\code{end}}{the end point of each segment;}
    \item{\code{A}}{an integer; the number of copies of the B allele
    at each segment;}
    \item{\code{B}}{an integer; the number of copies of the A allele
    at each segment;}
    \item{\code{parent.index}}{an integer indicating which clone
    each clone is descended from;}
    \item{\code{seg}}{integer; segment ID;}
  }

  The \code{dat} component is output from the function
  \code{snpDataGen}
}
\author{
  Kevin R. Coombes \email{krc@silicovore.com},
  Mark Zucker \email{zucker.64@buckeyemail.osu.edu}
}
\examples{
f.loci <- system.file("auxiliary/loci.Rda", package="TACG")
loci <- get(load(f.loci))
f.chlens <- system.file('auxiliary/chlens.Rda',package="TACG")
chlens <- get(load(f.chlens))
temp <- genSimChroms(N=3, chr=17, loci=loci, minLen=200000, maxLen=10000000, datapath=NULL, chlens=chlens, save=FALSE)
}
\keyword{}
